fix: move useTranslation out of SessionProviderWrapper to fix Mobile Safari crash
SessionProviderWrapper called useTranslation('settings') but rendered
outside I18nProvider in the layout tree. On Mobile Safari (iOS 18.7),
this caused a TypeError in react-i18next's internal useCallback when
the i18n context was missing.

Split the deep-link listener and translated error snackbar into a
separate NativeDeepLinkListener component that renders inside
I18nProvider where it has proper i18n and MUI theme context.
